backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.124.4",
    "httpx>=0.28.1",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
    "sqlalchemy>=2.0.45",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.ruff]
line-length = 79 


